+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
-
+### Added
+- default e-mail template, style and config
+- e-mail message for created shared user collections
 
 ## [2.32.1] - 2021-10-14
 ### Fixed